+++ npm/webpack-dev-server/src/util.ts
@@ -1,0 +1,7 @@
+import debug from 'debug'
+
+export const WBADebugNamespace = 'cypress-verbose:webpack-dev-server:bundle-analyzer'
+
+export const isWebpackBundleAnalyzerEnabled = () => {
+  return debug.enabled(WBADebugNamespace)
+}